+++ react/AddProductBtn.tsx
@@ -7,7 +7,7 @@ import React, {
   useEffect,
   SyntheticEvent,
 } from 'react'
-import PropTypes from 'prop-types'
+// import PropTypes from 'prop-types'
 import { useMutation, useLazyQuery } from 'react-apollo'
 import { defineMessages, useIntl } from 'react-intl'
 import { ProductContext } from 'vtex.product-context'
@@ -27,7 +27,7 @@ const localStore: any = storageFactory(() => sessionStorage)
 const CSS_HANDLES = ['wishlistIconContainer', 'wishlistIcon'] as const
 
 type AddBtnProps = {
-  toastURL: string
+  toastURL?: string
 }
 
 let isAuthenticated =
@@ -38,7 +38,7 @@ let wishListed: any =
   JSON.parse(localStore.getItem('wishlist_wishlisted')) ?? []
 
 const productCheck: {
-  [key: string]: { isWishlisted: boolean; wishListId: string }
+  [key: string]: { isWishlisted: boolean; wishListId: string; sku: string }
 } = {}
 const defaultValues = {
   LIST_NAME: 'Wishlist',
@@ -96,24 +96,30 @@ const useSessionResponse = () => {
   return session
 }
 
-const unify = (arr: any) => {
-  const obj: any = {}
-  for (let i = 0; i < arr.length; i++) {
-    obj[arr[i]] = true
+const addWishlisted = (productId: any, sku: any) => {
+  if (
+    wishListed.find(
+      (item: any) =>
+        item.productId &&
+        item.sku &&
+        item.productId === productId &&
+        item.sku === sku
+    ) === undefined
+  ) {
+    wishListed.push({
+      productId,
+      sku,
+    })
   }
-  return Object.getOwnPropertyNames(obj)
+  saveToLocalStorageItem(wishListed)
 }
 
-const addWishlisted = (productId: any) => {
-  if (wishListed.indexOf(productId) === -1) {
-    wishListed.push(productId)
-  }
-  wishListed = unify(wishListed)
-  localStore.setItem('wishlist_wishlisted', JSON.stringify(wishListed))
+const saveToLocalStorageItem = (data: any): any => {
+  localStore.setItem('wishlist_wishlisted', JSON.stringify(data))
+  return data
 }
 
-
-const AddBtn: FC<AddBtnProps> = ({ toastURL='/account/#wishlist' }) => {
+const AddBtn: FC<AddBtnProps> = ({ toastURL = '/account/#wishlist' }) => {
   const intl = useIntl()
   const [state, setState] = useState<any>({
     isLoading: true,
@@ -129,17 +135,14 @@ const AddBtn: FC<AddBtnProps> = ({ toastURL='/account/#wishlist' }) => {
           productCheck[productId] = {
             isWishlisted: false,
             wishListId: '',
+            sku: '',
           }
         }
 
-        wishListed = unify(wishListed)
-
-        const pos = wishListed.findIndex((item: string) => item === productId)
-
-        if (pos !== -1) {
-          wishListed.splice(pos, 1)
-          localStore.setItem('wishlist_wishlisted', JSON.stringify(wishListed))
-        }
+        wishListed = wishListed.filter(
+          (item: any) => !(item.productId === productId && item.sku === sku)
+        )
+        saveToLocalStorageItem(wishListed)
 
         setState({
           ...state,
@@ -157,6 +160,7 @@ const AddBtn: FC<AddBtnProps> = ({ toastURL='/account/#wishlist' }) => {
   const [handleCheck, { data, loading, called }] = useLazyQuery(checkItem)
 
   const [productId] = String(product?.productId).split('-')
+  const sku = product?.sku?.itemId
 
   const toastMessage = (messsageKey: string, linkWishlist: string) => {
     let action: any
@@ -196,8 +200,9 @@ const AddBtn: FC<AddBtnProps> = ({ toastURL='/account/#wishlist' }) => {
         productCheck[productId] = {
           wishListId: res.addToList,
           isWishlisted: true,
+          sku,
         }
-        addWishlisted(productId)
+        addWishlisted(productId, sku)
         toastMessage('productAddedToList', toastURL)
       },
     }
@@ -219,7 +224,7 @@ const AddBtn: FC<AddBtnProps> = ({ toastURL='/account/#wishlist' }) => {
     localStore.setItem('wishlist_shopperId', String(shopperId))
     if (!isAuthenticated && !shopperId) {
       if (localStore.getItem('wishlist_wishlisted')) {
-        localStore.removeItem('wishlist_wishlisted',)
+        localStore.removeItem('wishlist_wishlisted')
       }
     }
   }
@@ -261,17 +266,19 @@ const AddBtn: FC<AddBtnProps> = ({ toastURL='/account/#wishlist' }) => {
         variables: {
           shopperId: String(shopperId),
           productId,
+          sku,
         },
       })
     }
   }
+
   const checkFill = () => {
-    return (
-      sessionResponse?.namespaces?.profile?.isAuthenticated?.value === "false" ? false :
-      (wishListed.findIndex((item: string) => item === productId) !== -1 ||
-      productCheck[productId]?.isWishlisted ||
-      isWishlistPage)
-    )
+    return sessionResponse?.namespaces?.profile?.isAuthenticated?.value ===
+      'false'
+      ? false
+      : wishListed.find(
+          (item: any) => item.productId === productId && item.sku === sku
+        ) !== undefined
   }
 
   const handleAddProductClick = (e: SyntheticEvent) => {
@@ -283,8 +290,8 @@ const AddBtn: FC<AddBtnProps> = ({ toastURL='/account/#wishlist' }) => {
         items: {
           product,
           selectedItem,
-          account
-        }
+          account,
+        },
       }
 
       if (checkFill()) {
@@ -302,7 +309,7 @@ const AddBtn: FC<AddBtnProps> = ({ toastURL='/account/#wishlist' }) => {
             listItem: {
               productId,
               title: product.productName,
-              sku: selectedItem.itemId
+              sku: selectedItem.itemId,
             },
             shopperId,
             name: defaultValues.LIST_NAME,
@@ -310,11 +317,11 @@ const AddBtn: FC<AddBtnProps> = ({ toastURL='/account/#wishlist' }) => {
         })
         pixelEvent.event = 'addToWishlist'
       }
-      
+
       push(pixelEvent)
     } else {
       localStore.setItem('wishlist_addAfterLogin', String(productId))
-      toastMessage('notLogged',toastURL)
+      toastMessage('notLogged', toastURL)
     }
   }
 
@@ -330,18 +337,21 @@ const AddBtn: FC<AddBtnProps> = ({ toastURL='/account/#wishlist' }) => {
     productCheck[productId] = {
       isWishlisted: data.checkList.inList,
       wishListId: itemWishListId,
+      sku,
     }
 
     if (data.checkList.inList && wishListed.indexOf(productId) === -1) {
-      addWishlisted(productId)
+      addWishlisted(productId, sku)
     }
   } else if (
     data?.checkList?.inList === false &&
-    wishListed.length !== 0 && wishListed.indexOf(productId) !== -1
+    wishListed.length !== 0 &&
+    wishListed.indexOf(productId) !== -1
   ) {
     const indexWishListed = wishListed.indexOf(productId)
     wishListed.splice(indexWishListed, 1)
-    localStore.setItem('wishlist_wishlisted', JSON.stringify(wishListed))
+    // localStore.setItem('wishlist_wishlisted', JSON.stringify(wishListed))
+    saveToLocalStorageItem(wishListed)
   }
 
   return (
@@ -363,8 +373,8 @@ const AddBtn: FC<AddBtnProps> = ({ toastURL='/account/#wishlist' }) => {
   )
 }
 
-AddBtn.propTypes = {
-  toastURL: PropTypes.string.isRequired,
-}
+// AddBtn.propTypes = {
+//   toastURL: PropTypes.string.isRequired,
+// }
 
 export default AddBtn