+++ react/ProductSummaryWishlist.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable no-console */
 import React, { useMemo, useState, useEffect, FC } from 'react'
 import { useLazyQuery } from 'react-apollo'
 // @ts-expect-error - useTreePath is a private API
@@ -6,16 +7,16 @@ import { useListContext, ListContextProvider } from 'vtex.list-context'
 import { ProductListContext } from 'vtex.product-list-context'
 import { Spinner } from 'vtex.styleguide'
 import { usePixel } from 'vtex.pixel-manager'
+import { FormattedMessage } from 'react-intl'
 
 import { mapCatalogProductToProductSummary } from './utils/normalize'
 import ProductListEventCaller from './components/ProductListEventCaller'
 import productsQuery from './queries/productById.gql'
 import ViewLists from './queries/viewLists.gql'
 import { getSession } from './modules/session'
-import storageFactory from './utils/storage';
-import { FormattedMessage } from 'react-intl'
+import storageFactory from './utils/storage'
 
-const localStore = storageFactory(() => localStorage)
+const localStore = storageFactory(() => sessionStorage)
 
 let isAuthenticated =
   JSON.parse(String(localStore.getItem('wishlist_isAuthenticated'))) ?? false
@@ -41,13 +42,13 @@ const useSessionResponse = () => {
 }
 
 interface ProductSummaryProps {
-  children?: any,
+  children?: any
   showViewEmptyList?: boolean
 }
 
-const ProductSummaryList: FC<ProductSummaryProps> = ({ 
+const ProductSummaryList: FC<ProductSummaryProps> = ({
   children,
-  showViewEmptyList = false
+  showViewEmptyList = false,
 }) => {
   const { list } = useListContext() || []
   const { treePath } = useTreePath()
@@ -68,8 +69,11 @@ const ProductSummaryList: FC<ProductSummaryProps> = ({
     productsQuery,
     {
       ssr: false,
+      fetchPolicy: 'network-only',
     }
   )
+
+  console.log('loadProducts error =>', error)
 
   if (sessionResponse) {
     isAuthenticated =
@@ -89,14 +93,18 @@ const ProductSummaryList: FC<ProductSummaryProps> = ({
       })
     }
   }
+  let productList = [] as any
+  productList = dataLists?.viewLists[0]?.data.map((item: any) => {
+    const [id] = item.productId.split('-')
+    return {
+      productId: id,
+      sku: item.sku,
+    }
+  })
 
-  if (!called && dataLists) {
-    const ids = dataLists?.viewLists[0]?.data.map((item: any) => {
-      const [id] = item.productId.split('-')
-      return id
-    })
-
-    localStore.setItem('wishlist_wishlisted', JSON.stringify(ids))
+  if (!called && dataLists && productList) {
+    const ids = productList.map((item: any) => item.productId)
+    localStore.setItem('wishlist_wishlisted', JSON.stringify(productList))
     loadProducts({
       variables: {
         ids,
@@ -114,15 +122,37 @@ const ProductSummaryList: FC<ProductSummaryProps> = ({
         return itemId === id
       })?.id
     }
-    const componentList = products?.map((product: any, index: any) => {
+    let newProductList = []
+    if (productList) {
+      newProductList = productList
+        .map((_product: any) => {
+          if (products) {
+            const product = products.find(
+              (item: any) =>
+                item.productId === _product.productId &&
+                item.items &&
+                item.items.find((sku: any) => sku.itemId === _product.sku)
+            )
+            const _sku = product?.items.find(
+              (sku: any) => sku.itemId === _product.sku
+            )
+            return { ...product, sku: _sku }
+          }
+          return undefined
+        })
+        .filter((item: any) => item !== undefined)
+    }
+
+    const componentList = newProductList?.map((product: any, index: number) => {
       const sku = dataLists?.viewLists[0]?.data[index]?.sku
-      const items = data?.productsByIdentifier[index]?.items
+      const { items } = product
       const position = index + 1
 
       const normalizedProduct = mapCatalogProductToProductSummary(
         product,
         getWishlistId(product.productId)
       )
+
       if (sku && items.length) {
         for (const item of items) {
           if (item.itemId === sku) {
@@ -131,7 +161,7 @@ const ProductSummaryList: FC<ProductSummaryProps> = ({
         }
       }
 
-      const handleOnClick = () =>{
+      const handleOnClick = () => {
         push({
           event: 'productClick',
           list: 'wishlist',
@@ -151,7 +181,7 @@ const ProductSummaryList: FC<ProductSummaryProps> = ({
       )
     })
     return list.concat(componentList)
-  }, [products, treePath, list, dataLists])
+  }, [products, treePath, list, dataLists, productList, push])
 
   if (sessionResponse && !isAuthenticated) {
     navigate({
@@ -165,26 +195,17 @@ const ProductSummaryList: FC<ProductSummaryProps> = ({
   }
 
   if (!dataLists || !data || error) {
-    if (error && error?.message?.includes('products') && showViewEmptyList) {
-      return (
-        <ExtensionPoint
-          id="wishlist-empty-list"
-        />
-      )
+    if (error?.message?.includes('products') && showViewEmptyList) {
+      return <ExtensionPoint id="wishlist-empty-list" />
     }
     return null
   }
 
   if (listCalled && !listLoading && !dataLists?.viewLists[0]?.data?.length) {
     if (showViewEmptyList) {
-      return (
-        <ExtensionPoint
-          id="wishlist-empty-list"
-        />
-      )
-    } else {
-      return <FormattedMessage id="store/myaccount-empty-list" />
+      return <ExtensionPoint id="wishlist-empty-list" />
     }
+    return <FormattedMessage id="store/myaccount-empty-list" />
   }
 
   return (
@@ -195,7 +216,7 @@ const ProductSummaryList: FC<ProductSummaryProps> = ({
 }
 
 const EnhancedProductList: FC<ProductSummaryProps> = props => {
-  const { children, showViewEmptyList } = props;
+  const { children, showViewEmptyList } = props
   const { ProductListProvider } = ProductListContext
   return (
     <ProductListProvider listName="wishlist">